Fix namespace on unit tests for moved services
