+++ src/Akeneo/Pim/Automation/DataQualityInsights/tests/back/Specification/Application/ProductModelIdFactorySpec.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Specification\Akeneo\Pim\Automation\DataQualityInsights\Domain\ValueObject;
+namespace Specification\Akeneo\Pim\Automation\DataQualityInsights\Application;
 
 use Akeneo\Pim\Automation\DataQualityInsights\Domain\ValueObject\ProductIdCollection;
 use Akeneo\Pim\Automation\DataQualityInsights\Domain\ValueObject\ProductModelId;
